2004-04-02 18:01 UTC-0800 Ron Pinkas <ron@ronpinkas.com>   * include/hbapi.h     ! Corrected HB_IS*() to NOT use complex and wrong HB_IS_OF_TYPE()
  * include/hbfast.h
  * source/vm/fastitem.c
    - Removed redundant hb_fastClear()

  * source/compiler/hbpcode.c
    - Deffed OUT content of hb_compStrongType() - outdated!

  * source/rtl/filesys.c
    + Added temporary hack for xCC to provide _doserrno functionality.

  * source/vm/fastitem.c
  * source/vm/arrays.c
  * source/vm/memvars.c
  * source/vm/asort.c
  * source/vm/estack.c
  * source/vm/hvm.c
    + Added OWNERS support for HB_IT_BYREF

  * source/vm/fastitem.c
  * source/vm/arrays.c
  * source/vm/memvars.c
    ! Corrected HB_IS_*() to ->type == * (before reviewing and correcting HB_IS_*())

  * source/vm/codebloc.c
    ! Fixed GPF Trap in hb_codeblockDeleteGarbage()

  * source/vm/garbage.c
    * Converted static s_bReleaseAll to global hb_gc_bReleaseAll

  * source/vm/itemapi.c
    * Corrected compile warning

+++ xharbour/source/vm/itemapi.c
@@ -1,5 +1,5 @@
 /*
- * $Id: itemapi.c,v 1.90 2004/03/30 18:37:30 ronpinkas Exp $
+ * $Id: itemapi.c,v 1.91 2004/04/02 04:30:52 ronpinkas Exp $
  */
 
 /*
@@ -1244,7 +1244,7 @@ PHB_ITEM HB_EXPORT hb_itemUnRefOnce( PHB_ITEM pItem )
          {
             if( pItem->item.asRefer.offset == 0 )
             {
-               if( pItem->item.asRefer.BasePtr.pBaseArray->pItems == NULL || pItem->item.asRefer.BasePtr.pBaseArray->ulLen <= pItem->item.asRefer.value )
+               if( pItem->item.asRefer.BasePtr.pBaseArray->pItems == NULL || pItem->item.asRefer.BasePtr.pBaseArray->ulLen <= (ULONG) pItem->item.asRefer.value )
                {
                   HB_ITEM Array, Index;
